requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.52.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0.post1
nltk>=3.9.1
tqdm>=4.66.5
numpy>=1.26.4
sqlite-utils>=3.37
einops>=0.8.0
pytest>=8.3.3
watchdog>=4.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
jinja2>=3.1.3
python-multipart>=0.0.9
